Cleaning and Interpolation
Cleaned the main by moving the plotting functions in own scripts and into their own folder. Made changes to the interpolation in "getWindVec" so that the interpolation object is not called every time step but generated once and only the values are adapted. This way the internal grid remains the same, saving computational power.