Near field Bastankhah!
A lot of debugging! There was a major flaw in how the width of the field was calculated and plotted - that has changed. The updated equations are now fixed to +-3 sig_y and sig_z. Also the near field was added, which required bigger changes in the 'makeStep.m' file