Fixed a bug in initAtRotorPlane.m and updatePosition.m
The OPs did not spawn at the appropriate height since they were not initialized correctly. Also, they did not decrease in height. Reason was, that the bool was supposed to check if the OPs are under ground and reset them if neccessary. Instead of the negative position it was blocking a negative change.
Also assumed bug in the controller script: The CtCp factor seem to be stored in the wrong order (suggested by the var name) - the application of the values has been fixed and they look much more realistic now.